requirements.txt
Django==1.11.4
djangorestframework==3.6.3
ansible==1.8.4
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
pytz==2017.2
passlib==1.7.1
netifaces==0.10.4

